#17 Load mumbles from api
Ich habe hier mal ein wenig am TS gearbeitet und die User-Daten nur angezeigt, wenn man angemeldet ist. Die Card-Komponenten aus dem Design System braucht da noch ein Update, damit das übernommen wird.
Was mir aufgefallen ist: wenn man sich anmeldet, werden im ersten Moment keine Mumbles geladen. Bei dir auch? @THRY
