+++ Types/User.ts
@@ -1,0 +1,8 @@
+export type UserType = {
+  firstname?: string
+  lastname?: string
+  id: string
+  avatarUrl?: string
+  username?: string
+  email?: string
+}
